feat: connect everything to the right place
