+++ source/javascripts/pages/PipelinesPage/PipelinesPage.tsx
@@ -47,6 +47,7 @@ const EmptyState = () => {
   if (isGraphPipelinesEnabled) {
     return <CreateFirstGraphPipelineEmptyState onCreate={openDialog(PipelineConfigDialogType.CREATE_PIPELINE)} />;
   }
+
   return <StagePipelineEmptyState />;
 };
