+++ source/javascripts/pages/PipelinesPage/components/PipelineCanvas/GraphPipelineCanvas/components/WorkflowNode.tsx
@@ -1,18 +1,11 @@
 import { CSSProperties, useEffect, useRef } from 'react';
-import dagre from '@dagrejs/dagre';
 import { Box } from '@bitrise/bitkit';
 import { useResizeObserver } from 'usehooks-ts';
 import { Handle, Node, NodeProps, Position, useEdges, useReactFlow } from 'reactflow';
 import { PipelineWorkflow } from '@/core/models/Workflow';
 import { WorkflowCard } from '@/components/unified-editor';
-import {
-  CANVAS_PADDING,
-  TOOLBAR_CONTAINER_HEIGHT,
-  WORKFLOW_NODE_GAP_X,
-  WORKFLOW_NODE_GAP_Y,
-  WORKFLOW_NODE_HEIGHT,
-  WORKFLOW_NODE_WIDTH,
-} from '../GraphPipelineCanvas.const';
+import { WORKFLOW_NODE_HEIGHT, WORKFLOW_NODE_WIDTH } from '../GraphPipelineCanvas.const';
+import autoLayoutingGraphNodes from '../utils/autoLayoutingGraphNodes';
 
 type Props = NodeProps<PipelineWorkflow>;
 
@@ -26,33 +19,15 @@ const handleStyle: CSSProperties = {
 
 function updateNodeHeight(id: string, height?: number) {
   return (nodes: Node<PipelineWorkflow>[]) => {
-    const graph = new dagre.graphlib.Graph().setDefaultEdgeLabel(() => ({}));
+    return autoLayoutingGraphNodes(
+      nodes.map((n) => {
+        if (n.id !== id) {
+          return n;
+        }
 
-    graph.setGraph({
-      align: 'UL',
-      rankdir: 'LR',
-      marginx: CANVAS_PADDING - 160,
-      marginy: CANVAS_PADDING + TOOLBAR_CONTAINER_HEIGHT,
-      ranksep: WORKFLOW_NODE_GAP_X,
-      nodesep: WORKFLOW_NODE_GAP_Y,
-      edgesep: WORKFLOW_NODE_HEIGHT + WORKFLOW_NODE_GAP_Y,
-    });
-
-    nodes.forEach((node) => {
-      if (node.data.id === id) {
-        graph.setNode(node.data.id, { width: WORKFLOW_NODE_WIDTH, height: height ?? WORKFLOW_NODE_HEIGHT });
-      } else {
-        graph.setNode(node.data.id, { width: WORKFLOW_NODE_WIDTH, height: node.height ?? WORKFLOW_NODE_HEIGHT });
-      }
-      node.data.dependsOn.forEach((source) => graph.setEdge(source, node.id));
-    });
-
-    dagre.layout(graph);
-
-    return nodes.map((n) => {
-      const { x, y, height: h } = graph.node(n.id);
-      return { ...n, position: { x, y: y - h / 2 } };
-    });
+        return { ...n, height };
+      }),
+    );
   };
 }
