+++ source/javascripts/pages/PipelinesPage/components/PipelineCanvas/PipelineCanvas.tsx
@@ -14,13 +14,17 @@ const PipelineCanvas = () => {
   const CanvasComponent = variant === 'graph' ? GraphPipelineCanvas : StagedPipelineCanvas;
 
   return (
-    <Box bg="background/secondary" flex="1">
-      <Box position="relative" zIndex="50" p="16" mx="auto" maxW={800}>
-        <Toolbar
-          onPropertiesClick={openDialog(PipelineConfigDialogType.PIPELINE_CONFIG, selectedPipeline)}
-          onCreatePipelineClick={openDialog(PipelineConfigDialogType.CREATE_PIPELINE, '')}
-        />
-      </Box>
+    <Box bg="background/secondary" flex="1" position="relative">
+      <Toolbar
+        top="16"
+        left="50%"
+        zIndex="10"
+        position="absolute"
+        transform="translateX(-50%)"
+        width="clamp(0px, calc(100% - 32px), 768px)"
+        onPropertiesClick={openDialog(PipelineConfigDialogType.PIPELINE_CONFIG, selectedPipeline)}
+        onCreatePipelineClick={openDialog(PipelineConfigDialogType.CREATE_PIPELINE)}
+      />
       <CanvasComponent key={selectedPipeline} proOptions={{ hideAttribution: true }}>
         <Controls />
         <MiniMap />
